N253: Move annotations code out of cppmodel.
The (ffig-specific) annotations code has been moved out of cppmodel.py
and into annotations.py. This makes cppmodel more general (and
reusable).

Closes #253.
